+++ @WVStratification/placeParticlesOnIsopycnal.m
@@ -1,5 +1,5 @@
 function zIsopycnal = placeParticlesOnIsopycnal(wvt,x,y,zNoMotion)
-% places Lagrangian particles along a specified isopycnal
+% Return particle depths on the isopycnal identified by a no-motion depth.
 %
 % Given particle position (x,y), `zNoMotion` is used to determine the target
 % isopycnal using the no-motion density,
@@ -14,17 +14,17 @@ function zIsopycnal = placeParticlesOnIsopycnal(wvt,x,y,zNoMotion)
 % zIsopycnal = rho(targetRho);
 % ```
 %
-% where rho is the current total density field of the fluid.
+% where `rho` is the current total density field of the fluid.
 %
 % Note that the density is not necessarily monotonic, so the answer is not
 % necessarily unique.
 %
 % - Topic: Utilities
-% - Declaration: zIsopycnal = placeParticlesOnIsopycnal(x,y,z)
+% - Declaration: zIsopycnal = placeParticlesOnIsopycnal(x,y,zNoMotion)
 % - Parameter x: array of particle x positions
 % - Parameter y: array of particle y positions
-% - Parameter z: array of z location of target isopycnal in the no-motion profile
-% - Returns zIsopycnal: particle depth 
+% - Parameter zNoMotion: depths identifying target densities in the no-motion profile
+% - Returns zIsopycnal: particle depths in meters
 arguments
     wvt 
     x